requirements.txt
Django==4.1.3
django-allauth==0.51.0
django-crispy-forms==1.14.0
crispy-bootstrap5==0.7
django-allauth-ui==0.1.0
django-widget-tweaks==1.4.12
Pillow-9.3.0
django-colorfield-0.7.2